+++ pkg/__std/WriteArgs.go
@@ -42,7 +42,7 @@ func (rcv *WriteArgs) Value() []byte {
 	return nil
 }
 
-func (rcv *WriteArgs) Type() int8 {
+func (rcv *WriteArgs) Format() int8 {
 	o := flatbuffers.UOffsetT(rcv._tab.Offset(8))
 	if o != 0 {
 		return rcv._tab.GetInt8(o + rcv._tab.Pos)
@@ -50,7 +50,7 @@ func (rcv *WriteArgs) Type() int8 {
 	return 0
 }
 
-func (rcv *WriteArgs) MutateType(n int8) bool {
+func (rcv *WriteArgs) MutateFormat(n int8) bool {
 	return rcv._tab.MutateInt8Slot(8, n)
 }
 
@@ -75,8 +75,8 @@ func WriteArgsAddPath(builder *flatbuffers.Builder, path flatbuffers.UOffsetT) {
 func WriteArgsAddValue(builder *flatbuffers.Builder, value flatbuffers.UOffsetT) {
 	builder.PrependUOffsetTSlot(1, flatbuffers.UOffsetT(value), 0)
 }
-func WriteArgsAddType(builder *flatbuffers.Builder, type_ int8) {
-	builder.PrependInt8Slot(2, type_, 0)
+func WriteArgsAddFormat(builder *flatbuffers.Builder, format int8) {
+	builder.PrependInt8Slot(2, format, 0)
 }
 func WriteArgsAddIndent(builder *flatbuffers.Builder, indent int8) {
 	builder.PrependInt8Slot(3, indent, 0)
