--- std/__std_Deferred_generated.js
@@ -1,515 +1,0 @@
-// automatically generated by the FlatBuffers compiler, do not modify
-
-/**
- * @const
- * @namespace
- */
-var __std = __std || {};
-
-/**
- * @enum
- */
-__std.DeferredRetval = {
-  NONE: 0,
-  Deferred: 1,
-  Error: 2
-};
-
-/**
- * @enum
- */
-__std.FulfilmentValue = {
-  NONE: 0,
-  Data: 1,
-  EndOfStream: 2,
-  Error: 3
-};
-
-/**
- * CancelArgs encodes a message cancelling a deferred response.
- *
- * @constructor
- */
-__std.CancelArgs = function() {
-  /**
-   * @type {flatbuffers.ByteBuffer}
-   */
-  this.bb = null;
-
-  /**
-   * @type {number}
-   */
-  this.bb_pos = 0;
-};
-
-/**
- * @param {number} i
- * @param {flatbuffers.ByteBuffer} bb
- * @returns {__std.CancelArgs}
- */
-__std.CancelArgs.prototype.__init = function(i, bb) {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param {flatbuffers.ByteBuffer} bb
- * @param {__std.CancelArgs=} obj
- * @returns {__std.CancelArgs}
- */
-__std.CancelArgs.getRootAsCancelArgs = function(bb, obj) {
-  return (obj || new __std.CancelArgs).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @returns {flatbuffers.Long}
- */
-__std.CancelArgs.prototype.serial = function() {
-  var offset = this.bb.__offset(this.bb_pos, 4);
-  return offset ? this.bb.readUint64(this.bb_pos + offset) : this.bb.createLong(0, 0);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- */
-__std.CancelArgs.startCancelArgs = function(builder) {
-  builder.startObject(1);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {flatbuffers.Long} serial
- */
-__std.CancelArgs.addSerial = function(builder, serial) {
-  builder.addFieldInt64(0, serial, builder.createLong(0, 0));
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @returns {flatbuffers.Offset}
- */
-__std.CancelArgs.endCancelArgs = function(builder) {
-  var offset = builder.endObject();
-  return offset;
-};
-
-/**
- * Deferred encodes a serial number, standing in for an result or results to be fulfilled later.
- *
- * @constructor
- */
-__std.Deferred = function() {
-  /**
-   * @type {flatbuffers.ByteBuffer}
-   */
-  this.bb = null;
-
-  /**
-   * @type {number}
-   */
-  this.bb_pos = 0;
-};
-
-/**
- * @param {number} i
- * @param {flatbuffers.ByteBuffer} bb
- * @returns {__std.Deferred}
- */
-__std.Deferred.prototype.__init = function(i, bb) {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param {flatbuffers.ByteBuffer} bb
- * @param {__std.Deferred=} obj
- * @returns {__std.Deferred}
- */
-__std.Deferred.getRootAsDeferred = function(bb, obj) {
-  return (obj || new __std.Deferred).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @returns {flatbuffers.Long}
- */
-__std.Deferred.prototype.serial = function() {
-  var offset = this.bb.__offset(this.bb_pos, 4);
-  return offset ? this.bb.readUint64(this.bb_pos + offset) : this.bb.createLong(0, 0);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- */
-__std.Deferred.startDeferred = function(builder) {
-  builder.startObject(1);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {flatbuffers.Long} serial
- */
-__std.Deferred.addSerial = function(builder, serial) {
-  builder.addFieldInt64(0, serial, builder.createLong(0, 0));
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @returns {flatbuffers.Offset}
- */
-__std.Deferred.endDeferred = function(builder) {
-  var offset = builder.endObject();
-  return offset;
-};
-
-/**
- * Response encodes the _synchronous_ response to a message expecting a deferred value or values.
- *
- * @constructor
- */
-__std.DeferredResponse = function() {
-  /**
-   * @type {flatbuffers.ByteBuffer}
-   */
-  this.bb = null;
-
-  /**
-   * @type {number}
-   */
-  this.bb_pos = 0;
-};
-
-/**
- * @param {number} i
- * @param {flatbuffers.ByteBuffer} bb
- * @returns {__std.DeferredResponse}
- */
-__std.DeferredResponse.prototype.__init = function(i, bb) {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param {flatbuffers.ByteBuffer} bb
- * @param {__std.DeferredResponse=} obj
- * @returns {__std.DeferredResponse}
- */
-__std.DeferredResponse.getRootAsDeferredResponse = function(bb, obj) {
-  return (obj || new __std.DeferredResponse).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @returns {__std.DeferredRetval}
- */
-__std.DeferredResponse.prototype.retvalType = function() {
-  var offset = this.bb.__offset(this.bb_pos, 4);
-  return offset ? /** @type {__std.DeferredRetval} */ (this.bb.readUint8(this.bb_pos + offset)) : __std.DeferredRetval.NONE;
-};
-
-/**
- * @param {flatbuffers.Table} obj
- * @returns {?flatbuffers.Table}
- */
-__std.DeferredResponse.prototype.retval = function(obj) {
-  var offset = this.bb.__offset(this.bb_pos, 6);
-  return offset ? this.bb.__union(obj, this.bb_pos + offset) : null;
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- */
-__std.DeferredResponse.startDeferredResponse = function(builder) {
-  builder.startObject(2);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {__std.DeferredRetval} retvalType
- */
-__std.DeferredResponse.addRetvalType = function(builder, retvalType) {
-  builder.addFieldInt8(0, retvalType, __std.DeferredRetval.NONE);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {flatbuffers.Offset} retvalOffset
- */
-__std.DeferredResponse.addRetval = function(builder, retvalOffset) {
-  builder.addFieldOffset(1, retvalOffset, 0);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @returns {flatbuffers.Offset}
- */
-__std.DeferredResponse.endDeferredResponse = function(builder) {
-  var offset = builder.endObject();
-  return offset;
-};
-
-/**
- * Data is data sent as (possibly partial) fulfilment of a deferred result.
- *
- * @constructor
- */
-__std.Data = function() {
-  /**
-   * @type {flatbuffers.ByteBuffer}
-   */
-  this.bb = null;
-
-  /**
-   * @type {number}
-   */
-  this.bb_pos = 0;
-};
-
-/**
- * @param {number} i
- * @param {flatbuffers.ByteBuffer} bb
- * @returns {__std.Data}
- */
-__std.Data.prototype.__init = function(i, bb) {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param {flatbuffers.ByteBuffer} bb
- * @param {__std.Data=} obj
- * @returns {__std.Data}
- */
-__std.Data.getRootAsData = function(bb, obj) {
-  return (obj || new __std.Data).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @param {number} index
- * @returns {number}
- */
-__std.Data.prototype.bytes = function(index) {
-  var offset = this.bb.__offset(this.bb_pos, 4);
-  return offset ? this.bb.readUint8(this.bb.__vector(this.bb_pos + offset) + index) : 0;
-};
-
-/**
- * @returns {number}
- */
-__std.Data.prototype.bytesLength = function() {
-  var offset = this.bb.__offset(this.bb_pos, 4);
-  return offset ? this.bb.__vector_len(this.bb_pos + offset) : 0;
-};
-
-/**
- * @returns {Uint8Array}
- */
-__std.Data.prototype.bytesArray = function() {
-  var offset = this.bb.__offset(this.bb_pos, 4);
-  return offset ? new Uint8Array(this.bb.bytes().buffer, this.bb.bytes().byteOffset + this.bb.__vector(this.bb_pos + offset), this.bb.__vector_len(this.bb_pos + offset)) : null;
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- */
-__std.Data.startData = function(builder) {
-  builder.startObject(1);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {flatbuffers.Offset} bytesOffset
- */
-__std.Data.addBytes = function(builder, bytesOffset) {
-  builder.addFieldOffset(0, bytesOffset, 0);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {Array.<number>} data
- * @returns {flatbuffers.Offset}
- */
-__std.Data.createBytesVector = function(builder, data) {
-  builder.startVector(1, data.length, 1);
-  for (var i = data.length - 1; i >= 0; i--) {
-    builder.addInt8(data[i]);
-  }
-  return builder.endVector();
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {number} numElems
- */
-__std.Data.startBytesVector = function(builder, numElems) {
-  builder.startVector(1, numElems, 1);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @returns {flatbuffers.Offset}
- */
-__std.Data.endData = function(builder) {
-  var offset = builder.endObject();
-  return offset;
-};
-
-/**
- * EndOfStream signals that there will be no more data.
- *
- * @constructor
- */
-__std.EndOfStream = function() {
-  /**
-   * @type {flatbuffers.ByteBuffer}
-   */
-  this.bb = null;
-
-  /**
-   * @type {number}
-   */
-  this.bb_pos = 0;
-};
-
-/**
- * @param {number} i
- * @param {flatbuffers.ByteBuffer} bb
- * @returns {__std.EndOfStream}
- */
-__std.EndOfStream.prototype.__init = function(i, bb) {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param {flatbuffers.ByteBuffer} bb
- * @param {__std.EndOfStream=} obj
- * @returns {__std.EndOfStream}
- */
-__std.EndOfStream.getRootAsEndOfStream = function(bb, obj) {
-  return (obj || new __std.EndOfStream).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- */
-__std.EndOfStream.startEndOfStream = function(builder) {
-  builder.startObject(0);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @returns {flatbuffers.Offset}
- */
-__std.EndOfStream.endEndOfStream = function(builder) {
-  var offset = builder.endObject();
-  return offset;
-};
-
-/**
- * Resolution represents the _asynchronous_ fulfilment of a deferred value (possibly one of many values).
- *
- * @constructor
- */
-__std.Fulfilment = function() {
-  /**
-   * @type {flatbuffers.ByteBuffer}
-   */
-  this.bb = null;
-
-  /**
-   * @type {number}
-   */
-  this.bb_pos = 0;
-};
-
-/**
- * @param {number} i
- * @param {flatbuffers.ByteBuffer} bb
- * @returns {__std.Fulfilment}
- */
-__std.Fulfilment.prototype.__init = function(i, bb) {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param {flatbuffers.ByteBuffer} bb
- * @param {__std.Fulfilment=} obj
- * @returns {__std.Fulfilment}
- */
-__std.Fulfilment.getRootAsFulfilment = function(bb, obj) {
-  return (obj || new __std.Fulfilment).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @returns {flatbuffers.Long}
- */
-__std.Fulfilment.prototype.serial = function() {
-  var offset = this.bb.__offset(this.bb_pos, 4);
-  return offset ? this.bb.readUint64(this.bb_pos + offset) : this.bb.createLong(0, 0);
-};
-
-/**
- * @returns {__std.FulfilmentValue}
- */
-__std.Fulfilment.prototype.valueType = function() {
-  var offset = this.bb.__offset(this.bb_pos, 6);
-  return offset ? /** @type {__std.FulfilmentValue} */ (this.bb.readUint8(this.bb_pos + offset)) : __std.FulfilmentValue.NONE;
-};
-
-/**
- * @param {flatbuffers.Table} obj
- * @returns {?flatbuffers.Table}
- */
-__std.Fulfilment.prototype.value = function(obj) {
-  var offset = this.bb.__offset(this.bb_pos, 8);
-  return offset ? this.bb.__union(obj, this.bb_pos + offset) : null;
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- */
-__std.Fulfilment.startFulfilment = function(builder) {
-  builder.startObject(3);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {flatbuffers.Long} serial
- */
-__std.Fulfilment.addSerial = function(builder, serial) {
-  builder.addFieldInt64(0, serial, builder.createLong(0, 0));
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {__std.FulfilmentValue} valueType
- */
-__std.Fulfilment.addValueType = function(builder, valueType) {
-  builder.addFieldInt8(1, valueType, __std.FulfilmentValue.NONE);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {flatbuffers.Offset} valueOffset
- */
-__std.Fulfilment.addValue = function(builder, valueOffset) {
-  builder.addFieldOffset(2, valueOffset, 0);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @returns {flatbuffers.Offset}
- */
-__std.Fulfilment.endFulfilment = function(builder) {
-  var offset = builder.endObject();
-  return offset;
-};
-
-// Exports for ECMAScript6 Modules
-export {__std};
