--- std/__std_Error_generated.js
@@ -1,80 +1,0 @@
-// automatically generated by the FlatBuffers compiler, do not modify
-
-/**
- * @const
- * @namespace
- */
-var __std = __std || {};
-
-/**
- * Error encodes an error either in accepting a request, or in fulfilling it (in deferreds).
- *
- * @constructor
- */
-__std.Error = function() {
-  /**
-   * @type {flatbuffers.ByteBuffer}
-   */
-  this.bb = null;
-
-  /**
-   * @type {number}
-   */
-  this.bb_pos = 0;
-};
-
-/**
- * @param {number} i
- * @param {flatbuffers.ByteBuffer} bb
- * @returns {__std.Error}
- */
-__std.Error.prototype.__init = function(i, bb) {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param {flatbuffers.ByteBuffer} bb
- * @param {__std.Error=} obj
- * @returns {__std.Error}
- */
-__std.Error.getRootAsError = function(bb, obj) {
-  return (obj || new __std.Error).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @param {flatbuffers.Encoding=} optionalEncoding
- * @returns {string|Uint8Array|null}
- */
-__std.Error.prototype.message = function(optionalEncoding) {
-  var offset = this.bb.__offset(this.bb_pos, 4);
-  return offset ? this.bb.__string(this.bb_pos + offset, optionalEncoding) : null;
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- */
-__std.Error.startError = function(builder) {
-  builder.startObject(1);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {flatbuffers.Offset} messageOffset
- */
-__std.Error.addMessage = function(builder, messageOffset) {
-  builder.addFieldOffset(0, messageOffset, 0);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @returns {flatbuffers.Offset}
- */
-__std.Error.endError = function(builder) {
-  var offset = builder.endObject();
-  return offset;
-};
-
-// Exports for ECMAScript6 Modules
-export {__std};
