--- std/__std_FileSystem_generated.js
@@ -1,473 +1,0 @@
-// automatically generated by the FlatBuffers compiler, do not modify
-
-/**
- * @const
- * @namespace
- */
-var __std = __std || {};
-
-/**
- * @enum
- */
-__std.FileSystemRetval = {
-  NONE: 0,
-  FileInfo: 1,
-  Directory: 2,
-  Error: 3
-};
-
-/**
- * @constructor
- */
-__std.FileInfo = function() {
-  /**
-   * @type {flatbuffers.ByteBuffer}
-   */
-  this.bb = null;
-
-  /**
-   * @type {number}
-   */
-  this.bb_pos = 0;
-};
-
-/**
- * @param {number} i
- * @param {flatbuffers.ByteBuffer} bb
- * @returns {__std.FileInfo}
- */
-__std.FileInfo.prototype.__init = function(i, bb) {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param {flatbuffers.ByteBuffer} bb
- * @param {__std.FileInfo=} obj
- * @returns {__std.FileInfo}
- */
-__std.FileInfo.getRootAsFileInfo = function(bb, obj) {
-  return (obj || new __std.FileInfo).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @param {flatbuffers.Encoding=} optionalEncoding
- * @returns {string|Uint8Array|null}
- */
-__std.FileInfo.prototype.name = function(optionalEncoding) {
-  var offset = this.bb.__offset(this.bb_pos, 4);
-  return offset ? this.bb.__string(this.bb_pos + offset, optionalEncoding) : null;
-};
-
-/**
- * @param {flatbuffers.Encoding=} optionalEncoding
- * @returns {string|Uint8Array|null}
- */
-__std.FileInfo.prototype.path = function(optionalEncoding) {
-  var offset = this.bb.__offset(this.bb_pos, 6);
-  return offset ? this.bb.__string(this.bb_pos + offset, optionalEncoding) : null;
-};
-
-/**
- * @returns {boolean}
- */
-__std.FileInfo.prototype.isdir = function() {
-  var offset = this.bb.__offset(this.bb_pos, 8);
-  return offset ? !!this.bb.readInt8(this.bb_pos + offset) : false;
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- */
-__std.FileInfo.startFileInfo = function(builder) {
-  builder.startObject(3);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {flatbuffers.Offset} nameOffset
- */
-__std.FileInfo.addName = function(builder, nameOffset) {
-  builder.addFieldOffset(0, nameOffset, 0);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {flatbuffers.Offset} pathOffset
- */
-__std.FileInfo.addPath = function(builder, pathOffset) {
-  builder.addFieldOffset(1, pathOffset, 0);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {boolean} isdir
- */
-__std.FileInfo.addIsdir = function(builder, isdir) {
-  builder.addFieldInt8(2, +isdir, +false);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @returns {flatbuffers.Offset}
- */
-__std.FileInfo.endFileInfo = function(builder) {
-  var offset = builder.endObject();
-  return offset;
-};
-
-/**
- * @constructor
- */
-__std.Directory = function() {
-  /**
-   * @type {flatbuffers.ByteBuffer}
-   */
-  this.bb = null;
-
-  /**
-   * @type {number}
-   */
-  this.bb_pos = 0;
-};
-
-/**
- * @param {number} i
- * @param {flatbuffers.ByteBuffer} bb
- * @returns {__std.Directory}
- */
-__std.Directory.prototype.__init = function(i, bb) {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param {flatbuffers.ByteBuffer} bb
- * @param {__std.Directory=} obj
- * @returns {__std.Directory}
- */
-__std.Directory.getRootAsDirectory = function(bb, obj) {
-  return (obj || new __std.Directory).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @param {flatbuffers.Encoding=} optionalEncoding
- * @returns {string|Uint8Array|null}
- */
-__std.Directory.prototype.name = function(optionalEncoding) {
-  var offset = this.bb.__offset(this.bb_pos, 4);
-  return offset ? this.bb.__string(this.bb_pos + offset, optionalEncoding) : null;
-};
-
-/**
- * @param {flatbuffers.Encoding=} optionalEncoding
- * @returns {string|Uint8Array|null}
- */
-__std.Directory.prototype.path = function(optionalEncoding) {
-  var offset = this.bb.__offset(this.bb_pos, 6);
-  return offset ? this.bb.__string(this.bb_pos + offset, optionalEncoding) : null;
-};
-
-/**
- * @param {number} index
- * @param {__std.FileInfo=} obj
- * @returns {__std.FileInfo}
- */
-__std.Directory.prototype.files = function(index, obj) {
-  var offset = this.bb.__offset(this.bb_pos, 8);
-  return offset ? (obj || new __std.FileInfo).__init(this.bb.__indirect(this.bb.__vector(this.bb_pos + offset) + index * 4), this.bb) : null;
-};
-
-/**
- * @returns {number}
- */
-__std.Directory.prototype.filesLength = function() {
-  var offset = this.bb.__offset(this.bb_pos, 8);
-  return offset ? this.bb.__vector_len(this.bb_pos + offset) : 0;
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- */
-__std.Directory.startDirectory = function(builder) {
-  builder.startObject(3);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {flatbuffers.Offset} nameOffset
- */
-__std.Directory.addName = function(builder, nameOffset) {
-  builder.addFieldOffset(0, nameOffset, 0);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {flatbuffers.Offset} pathOffset
- */
-__std.Directory.addPath = function(builder, pathOffset) {
-  builder.addFieldOffset(1, pathOffset, 0);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {flatbuffers.Offset} filesOffset
- */
-__std.Directory.addFiles = function(builder, filesOffset) {
-  builder.addFieldOffset(2, filesOffset, 0);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {Array.<flatbuffers.Offset>} data
- * @returns {flatbuffers.Offset}
- */
-__std.Directory.createFilesVector = function(builder, data) {
-  builder.startVector(4, data.length, 4);
-  for (var i = data.length - 1; i >= 0; i--) {
-    builder.addOffset(data[i]);
-  }
-  return builder.endVector();
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {number} numElems
- */
-__std.Directory.startFilesVector = function(builder, numElems) {
-  builder.startVector(4, numElems, 4);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @returns {flatbuffers.Offset}
- */
-__std.Directory.endDirectory = function(builder) {
-  var offset = builder.endObject();
-  return offset;
-};
-
-/**
- * @constructor
- */
-__std.FileInfoArgs = function() {
-  /**
-   * @type {flatbuffers.ByteBuffer}
-   */
-  this.bb = null;
-
-  /**
-   * @type {number}
-   */
-  this.bb_pos = 0;
-};
-
-/**
- * @param {number} i
- * @param {flatbuffers.ByteBuffer} bb
- * @returns {__std.FileInfoArgs}
- */
-__std.FileInfoArgs.prototype.__init = function(i, bb) {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param {flatbuffers.ByteBuffer} bb
- * @param {__std.FileInfoArgs=} obj
- * @returns {__std.FileInfoArgs}
- */
-__std.FileInfoArgs.getRootAsFileInfoArgs = function(bb, obj) {
-  return (obj || new __std.FileInfoArgs).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @param {flatbuffers.Encoding=} optionalEncoding
- * @returns {string|Uint8Array|null}
- */
-__std.FileInfoArgs.prototype.path = function(optionalEncoding) {
-  var offset = this.bb.__offset(this.bb_pos, 4);
-  return offset ? this.bb.__string(this.bb_pos + offset, optionalEncoding) : null;
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- */
-__std.FileInfoArgs.startFileInfoArgs = function(builder) {
-  builder.startObject(1);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {flatbuffers.Offset} pathOffset
- */
-__std.FileInfoArgs.addPath = function(builder, pathOffset) {
-  builder.addFieldOffset(0, pathOffset, 0);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @returns {flatbuffers.Offset}
- */
-__std.FileInfoArgs.endFileInfoArgs = function(builder) {
-  var offset = builder.endObject();
-  return offset;
-};
-
-/**
- * @constructor
- */
-__std.ListArgs = function() {
-  /**
-   * @type {flatbuffers.ByteBuffer}
-   */
-  this.bb = null;
-
-  /**
-   * @type {number}
-   */
-  this.bb_pos = 0;
-};
-
-/**
- * @param {number} i
- * @param {flatbuffers.ByteBuffer} bb
- * @returns {__std.ListArgs}
- */
-__std.ListArgs.prototype.__init = function(i, bb) {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param {flatbuffers.ByteBuffer} bb
- * @param {__std.ListArgs=} obj
- * @returns {__std.ListArgs}
- */
-__std.ListArgs.getRootAsListArgs = function(bb, obj) {
-  return (obj || new __std.ListArgs).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @param {flatbuffers.Encoding=} optionalEncoding
- * @returns {string|Uint8Array|null}
- */
-__std.ListArgs.prototype.path = function(optionalEncoding) {
-  var offset = this.bb.__offset(this.bb_pos, 4);
-  return offset ? this.bb.__string(this.bb_pos + offset, optionalEncoding) : null;
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- */
-__std.ListArgs.startListArgs = function(builder) {
-  builder.startObject(1);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {flatbuffers.Offset} pathOffset
- */
-__std.ListArgs.addPath = function(builder, pathOffset) {
-  builder.addFieldOffset(0, pathOffset, 0);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @returns {flatbuffers.Offset}
- */
-__std.ListArgs.endListArgs = function(builder) {
-  var offset = builder.endObject();
-  return offset;
-};
-
-/**
- * @constructor
- */
-__std.FileSystemResponse = function() {
-  /**
-   * @type {flatbuffers.ByteBuffer}
-   */
-  this.bb = null;
-
-  /**
-   * @type {number}
-   */
-  this.bb_pos = 0;
-};
-
-/**
- * @param {number} i
- * @param {flatbuffers.ByteBuffer} bb
- * @returns {__std.FileSystemResponse}
- */
-__std.FileSystemResponse.prototype.__init = function(i, bb) {
-  this.bb_pos = i;
-  this.bb = bb;
-  return this;
-};
-
-/**
- * @param {flatbuffers.ByteBuffer} bb
- * @param {__std.FileSystemResponse=} obj
- * @returns {__std.FileSystemResponse}
- */
-__std.FileSystemResponse.getRootAsFileSystemResponse = function(bb, obj) {
-  return (obj || new __std.FileSystemResponse).__init(bb.readInt32(bb.position()) + bb.position(), bb);
-};
-
-/**
- * @returns {__std.FileSystemRetval}
- */
-__std.FileSystemResponse.prototype.retvalType = function() {
-  var offset = this.bb.__offset(this.bb_pos, 4);
-  return offset ? /** @type {__std.FileSystemRetval} */ (this.bb.readUint8(this.bb_pos + offset)) : __std.FileSystemRetval.NONE;
-};
-
-/**
- * @param {flatbuffers.Table} obj
- * @returns {?flatbuffers.Table}
- */
-__std.FileSystemResponse.prototype.retval = function(obj) {
-  var offset = this.bb.__offset(this.bb_pos, 6);
-  return offset ? this.bb.__union(obj, this.bb_pos + offset) : null;
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- */
-__std.FileSystemResponse.startFileSystemResponse = function(builder) {
-  builder.startObject(2);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {__std.FileSystemRetval} retvalType
- */
-__std.FileSystemResponse.addRetvalType = function(builder, retvalType) {
-  builder.addFieldInt8(0, retvalType, __std.FileSystemRetval.NONE);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @param {flatbuffers.Offset} retvalOffset
- */
-__std.FileSystemResponse.addRetval = function(builder, retvalOffset) {
-  builder.addFieldOffset(1, retvalOffset, 0);
-};
-
-/**
- * @param {flatbuffers.Builder} builder
- * @returns {flatbuffers.Offset}
- */
-__std.FileSystemResponse.endFileSystemResponse = function(builder) {
-  var offset = builder.endObject();
-  return offset;
-};
-
-// Exports for ECMAScript6 Modules
-export {__std};
