+++ std/__std_Format.fbs
@@ -1,0 +1,8 @@
+namespace __std;
+
+enum Format : byte {
+    Auto,
+    JSON,
+    YAML,
+    Raw,
+}
