+++ std/generate.sh
@@ -4,6 +4,7 @@ std=$(dirname $0)
 root=${std}/..
 
 for f in `ls ${std}/*.fbs`; do
-  flatc --js --es6-js-export -o ${std} ${f}
   flatc --go -o ${root}/pkg ${f}
 done
+
+flatc --js --gen-onefile --gen-all --es6-js-export -o ${std} ${std}/__std.fbs
