+++ std/std_write.js
@@ -1,8 +1,7 @@
-import { __std as w } from '__std_Write_generated';
-import { __std as m } from '__std_generated';
+import { __std } from '__std_generated';
 import flatbuffers from 'flatbuffers';
 
-const Format = Object.freeze(w.Format);
+const Format = Object.freeze(__std.Format);
 
 function write(value, path = '', { format = Format.Auto, indent = 2 } = {}) {
   const builder = new flatbuffers.Builder(1024);
@@ -10,17 +9,17 @@ function write(value, path = '', { format = Format.Auto, indent = 2 } = {}) {
   const strOff = builder.createString(str);
   const pathOff = builder.createString(path);
 
-  w.WriteArgs.startWriteArgs(builder);
-  w.WriteArgs.addValue(builder, strOff);
-  w.WriteArgs.addPath(builder, pathOff);
-  w.WriteArgs.addType(builder, format);
-  w.WriteArgs.addIndent(builder, indent);
-  const args = w.WriteArgs.endWriteArgs(builder);
+  __std.WriteArgs.startWriteArgs(builder);
+  __std.WriteArgs.addValue(builder, strOff);
+  __std.WriteArgs.addPath(builder, pathOff);
+  __std.WriteArgs.addFormat(builder, format);
+  __std.WriteArgs.addIndent(builder, indent);
+  const args = __std.WriteArgs.endWriteArgs(builder);
 
-  m.Message.startMessage(builder);
-  m.Message.addArgsType(builder, m.Args.WriteArgs);
-  m.Message.addArgs(builder, args);
-  const message = m.Message.endMessage(builder);
+  __std.Message.startMessage(builder);
+  __std.Message.addArgsType(builder, __std.Args.WriteArgs);
+  __std.Message.addArgs(builder, args);
+  const message = __std.Message.endMessage(builder);
 
   builder.finish(message);
   V8Worker2.send(builder.asArrayBuffer());
